+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * include the argument `which_plots` in the `plot.lcc` function so that one can
   specify which plots should be plotted
+* include the argument `n_item_consensus` in the `plot.lcc` function so that one
+  can specify how many item consensus plots should be plotted in one plot
 
 # longmixr 1.0.0